🚚 package 구조 변경
- 크게 global과 domain으로 나누었다
