+++ src/main/java/com/jpaplayground/domain/conversation/UserType.java
@@ -1,4 +1,4 @@
-package com.jpaplayground.conversation;
+package com.jpaplayground.domain.conversation;
 
 public enum UserType {
 	SELLER,
