+++ src/main/java/com/jpaplayground/domain/product/Product.java
@@ -1,4 +1,4 @@
-package com.jpaplayground.product;
+package com.jpaplayground.domain.product;
 
 import javax.persistence.Entity;
 import javax.persistence.GeneratedValue;
