+++ src/main/java/com/jpaplayground/domain/product/ProductService.java
@@ -1,7 +1,7 @@
-package com.jpaplayground.product;
+package com.jpaplayground.domain.product;
 
-import com.jpaplayground.product.dto.ProductCreateRequest;
-import com.jpaplayground.product.dto.ProductResponse;
+import com.jpaplayground.domain.product.dto.ProductCreateRequest;
+import com.jpaplayground.domain.product.dto.ProductResponse;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.domain.Slice;
